GUI-258. MyProfilesStotage: Move currentProfile details from OrderCreateStore to MyProfilesStore.
